+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 ~~~~~
 * Migration to convert `author` roles to `translator` when the `Resource` is a
   translation.
+* Field to record fictional places of publication.
 
 Changed
 ~~~~~~~